--- src/cascader/utils/panel.ts
@@ -1,122 +1,0 @@
-import isNumber from 'lodash/isNumber';
-import { TreeNode, CascaderContextType, CascaderProps, TreeNodeValue } from '../interface';
-
-/**
- * 面板数据计算方法
- * @param treeNodes
- * @returns
- */
-export function getPanels(treeNodes: CascaderContextType['treeNodes']) {
-  const panels: TreeNode[][] = [];
-  treeNodes.forEach((node: TreeNode) => {
-    if (panels[node.level]) {
-      panels[node.level].push(node);
-    } else {
-      panels[node.level] = [node];
-    }
-  });
-  return panels;
-}
-
-/**
- * 点击item的副作用
- * @param propsTrigger
- * @param trigger
- * @param node
- * @param cascaderContext
- * @param onChange
- * @param ctx
- */
-export function expendClickEffect(
-  propsTrigger: CascaderProps['trigger'],
-  trigger: CascaderProps['trigger'],
-  node: TreeNode,
-  cascaderContext: CascaderContextType,
-) {
-  const { checkStrictly, multiple, treeStore, setVisible, setValue, setTreeNodes, setExpend, value, max, valueType } =
-    cascaderContext;
-
-  const isDisabled = node.disabled || (multiple && (value as TreeNodeValue[]).length >= max && max !== 0);
-
-  if (isDisabled) return;
-  // 点击展开节点，设置展开状态
-  if (propsTrigger === trigger && !node.isLeaf()) {
-    const expanded = node.setExpanded(true);
-    treeStore.refreshNodes();
-    treeStore.replaceExpanded(expanded);
-    const nodes = treeStore.getNodes().filter((node: TreeNode) => node.visible);
-    setTreeNodes(nodes);
-
-    // 多选条件下手动维护expend
-    if (multiple) {
-      setExpend(expanded);
-    }
-  }
-
-  if (!multiple && (node.isLeaf() || checkStrictly) && trigger === 'click') {
-    treeStore.resetChecked();
-    const checked = node.setChecked(!node.isChecked());
-    const [value] = checked;
-
-    // 非受控状态下更新状态
-    setValue(valueType === 'single' ? value : node.getPath().map((item) => item.value), 'checked', node.getModel());
-
-    if (!checkStrictly) {
-      setVisible(false);
-    }
-  }
-}
-
-/**
- * 多选状态下选中状态数据变化的副作用
- * @param node
- * @param cascaderContext
- * @returns
- */
-export function valueChangeEffect(node: TreeNode, cascaderContext: CascaderContextType) {
-  const { disabled, max, inputVal, multiple, setVisible, setValue, treeNodes, treeStore, valueType } = cascaderContext;
-
-  if (!node || disabled || node.disabled) {
-    return;
-  }
-  const checked = node.setChecked(!node.isChecked());
-
-  if (isNumber(max) && max < 0) {
-    console.warn('TDesign Warn:', 'max should > 0');
-  }
-
-  if (checked.length > max && isNumber(max) && max > 0) {
-    return;
-  }
-
-  if (checked.length === 0) {
-    const expanded = treeStore.getExpanded();
-    setTimeout(() => {
-      treeStore.replaceExpanded(expanded);
-      treeStore.refreshNodes();
-    }, 0);
-  }
-
-  if (!multiple) {
-    setVisible(false);
-  }
-
-  const isSelectAll = treeNodes.every((item) => checked.indexOf(item.value) > -1);
-
-  if (inputVal && isSelectAll) {
-    setVisible(false);
-  }
-
-  // 处理不同数据类型
-  const resValue =
-    valueType === 'single'
-      ? checked
-      : checked.map((val) =>
-          treeStore
-            .getNode(val)
-            .getPath()
-            .map((item) => item.value),
-        );
-
-  setValue(resValue, 'checked', node.getModel());
-}
